Move BuildFrom and BuildTo from the tests into the main project
